// src/types/followerList/followers.d.ts
// Generated by https://quicktype.io

export interface Follows {
  message: string;
  code: number;
  data: FollowData;
}

export interface FollowData {
  minCreatedAt: string;
  users: User[];
}

export interface User {
  id: number;
  nickname: string;
  imageURL: string;
  title: string;
  follow: boolean;
}